Basic Setup
There is a stat manager which controls the data and sends out signals when stats change. The screen is divided into a top row where there are labels for each of the stats and a main area where all actions will exist. Actions are children of a grid container. Each action consists of a label which tells you what it does and a button which does it.